Enable multi-threading forwarding
The forward_proxy is only accepting one request at a time, which is particularly unsuitable if the server is not responding or we have intermittent links. For example, if for some reason the first ACK sent be the proxy is lost, no other ACKs are sent (until the response), and the client might give up if the server is taking some time to respond.