Lost ACKs and message deduplication
When using the threaded-version forward_proxy it might occur that the first ACK (sent by the proxy is lost) and the client will retransmit the request. This leads to a duplicate message detection by the proxy which in turn sends an empty message with the new CON identifier. Without this fix the client will receive the empty message that corresponds to an existing transaction and stop the retransmission without acknowledging this transaction.

From the RFC there are several approaches to handle these duplicates (https://tools.ietf.org/html/rfc7252#section-4.5), I opted to implicitly acknowledge the transaction when receiving the CON message to avoid sending another ACK.